Ensure m2e 1.8.2 works with older jdt versions
