+++ org.eclipse.m2e.jdt/src/org/eclipse/m2e/jdt/internal/ModuleSupport.java
@@ -11,39 +11,11 @@
 
 package org.eclipse.m2e.jdt.internal;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.zip.ZipFile;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import org.eclipse.core.resources.IProject;
-import org.eclipse.core.resources.IWorkspaceRoot;
-import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.jdt.core.IClasspathAttribute;
-import org.eclipse.jdt.core.IClasspathEntry;
-import org.eclipse.jdt.core.IJavaProject;
-import org.eclipse.jdt.core.IModuleDescription;
-import org.eclipse.jdt.core.JavaCore;
-import org.eclipse.jdt.internal.compiler.classfmt.ClassFileReader;
-import org.eclipse.jdt.internal.compiler.classfmt.ClassFormatException;
-import org.eclipse.jdt.internal.compiler.env.AutomaticModuleNaming;
-import org.eclipse.jdt.internal.compiler.env.IModule;
-import org.eclipse.jdt.internal.core.AbstractModule;
 
-import org.apache.maven.project.MavenProject;
-
-import org.eclipse.m2e.core.MavenPlugin;
 import org.eclipse.m2e.core.project.IMavenProjectFacade;
 import org.eclipse.m2e.jdt.IClasspathDescriptor;
-import org.eclipse.m2e.jdt.IClasspathEntryDescriptor;
 
 
 /**
@@ -52,12 +24,9 @@ import org.eclipse.m2e.jdt.IClasspathEntryDescriptor;
  * @author Fred Bricon
  * @since 1.8.2
  */
-@SuppressWarnings("restriction")
 public class ModuleSupport {
 
   static final boolean IS_MODULE_SUPPORT_AVAILABLE;
-
-  private static final Logger log = LoggerFactory.getLogger(ModuleSupport.class);
 
   static {
     boolean isModuleSupportAvailable = false;
@@ -81,93 +50,6 @@ public class ModuleSupport {
     if(!IS_MODULE_SUPPORT_AVAILABLE) {
       return;
     }
-    IJavaProject javaProject = JavaCore.create(facade.getProject());
-    IModuleDescription moduleDescription = javaProject.getModuleDescription();
-    if(!(moduleDescription instanceof AbstractModule)) {
-      return;
-    }
-    AbstractModule module = (AbstractModule) moduleDescription;
-    Set<String> requiredModules = Stream.of(module.getRequiredModules()).map(m -> new String(m.name()))
-        .collect(Collectors.toSet());
-    for(IClasspathEntryDescriptor entry : classpath.getEntryDescriptors()) {
-      String moduleName = getModuleName(entry, monitor);
-      if(requiredModules.contains(moduleName)) {
-        entry.setClasspathAttribute(IClasspathAttribute.MODULE, Boolean.TRUE.toString());
-      }
-    }
+    InternalModuleSupport.configureClasspath(facade, classpath, monitor);
   }
-
-  private static String getModuleName(IClasspathEntryDescriptor entry, IProgressMonitor monitor) {
-    String module = null;
-    if(IClasspathEntry.CPE_LIBRARY == entry.getEntryKind()) {
-      module = getModuleNameFromJar(entry.getPath().toFile());
-    } else if(IClasspathEntry.CPE_PROJECT == entry.getEntryKind()) {
-      module = getModuleNameFromProject(entry.getPath(), monitor);
-    }
-    return module;
-  }
-
-  private static String getModuleNameFromProject(IPath projectPath, IProgressMonitor monitor) {
-    IJavaProject project = getJavaProject(projectPath);
-    String module = null;
-    if(project != null) {
-      try {
-        if(project.getModuleDescription() == null) {
-          String buildName = null;
-          IMavenProjectFacade facade = MavenPlugin.getMavenProjectRegistry().getProject(project.getProject());
-          if(facade != null) {
-            MavenProject mavenProject = facade.getMavenProject(monitor);
-            if(mavenProject != null) {
-              buildName = mavenProject.getBuild().getFinalName();
-            }
-          }
-          if(buildName == null || buildName.isEmpty()) {
-            buildName = project.getElementName();
-          }
-          module = new String(AutomaticModuleNaming.determineAutomaticModuleName(buildName, false, null));
-        } else {
-          module = project.getModuleDescription().getElementName();
-        }
-      } catch(CoreException ex) {
-        log.error(ex.getMessage(), ex);
-      }
-    }
-    return module;
-  }
-
-  private static IJavaProject getJavaProject(IPath projectPath) {
-    if(projectPath == null || projectPath.isEmpty()) {
-      return null;
-    }
-    IWorkspaceRoot root = ResourcesPlugin.getWorkspace().getRoot();
-    IProject project = root.getProject(projectPath.lastSegment());
-    if(project.isAccessible()) {
-      return JavaCore.create(project);
-    }
-    return null;
-  }
-
-  private static String getModuleNameFromJar(File file) {
-    if(!file.isFile()) {
-      return null;
-    }
-    char[] moduleName = null;
-    try (ZipFile zipFile = new ZipFile(file)) {
-      IModule module = null;
-      ClassFileReader reader = ClassFileReader.read(zipFile, IModule.MODULE_INFO_CLASS);
-      if(reader != null) {
-        module = reader.getModuleDeclaration();
-        if(module != null) {
-          moduleName = module.name();
-        }
-      }
-    } catch(ClassFormatException | IOException ex) {
-      log.error(ex.getMessage(), ex);
-    }
-    if(moduleName == null) {
-      moduleName = AutomaticModuleNaming.determineAutomaticModuleName(file.getAbsolutePath());
-    }
-    return new String(moduleName);
-  }
-
 }
